requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.3.0
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
idna==3.6
protobuf==3.20.3
pydantic==2.5.3
pydantic_core==2.14.6
PyYAML==6.0.1
requests==2.31.0
setuptools==75.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
